Use raw data storage for events to collect more info
OSprofiler uses free-form notifications in fact, so any kind of data
might be included to the profiling trace. To be able to do that in
Ceilometer case, that use explicit event definitions, we need to
use its raw data events storage, that will include all notification
message.

To use this feature the following section needs to be set in
ceilometer.conf:

[event]
store_raw = info

where info is the notification level used.

Warning: if raw data is used, ALL events that come into ceilometer
with info notification level will be stored with raw data included,
there is no way to set this parameter only for profiler.* events.

Change-Id: I3ebf8ee34ac01e68d530a676ae840f7a62d897bd